+++ packages/effect/src/internal/cause.ts
@@ -686,7 +686,7 @@ const evaluateCause = (
         break
       }
       case OpCodes.OP_FAIL: {
-        _parallel = HashSet.add(_parallel, cause.error)
+        _parallel = HashSet.add(_parallel, Chunk.make(cause._tag, cause.error))
         if (stack.length === 0) {
           return [_parallel, _sequential]
         }
@@ -694,7 +694,7 @@ const evaluateCause = (
         break
       }
       case OpCodes.OP_DIE: {
-        _parallel = HashSet.add(_parallel, cause.defect)
+        _parallel = HashSet.add(_parallel, Chunk.make(cause._tag, cause.defect))
         if (stack.length === 0) {
           return [_parallel, _sequential]
         }
@@ -702,7 +702,7 @@ const evaluateCause = (
         break
       }
       case OpCodes.OP_INTERRUPT: {
-        _parallel = HashSet.add(_parallel, cause.fiberId as unknown)
+        _parallel = HashSet.add(_parallel, Chunk.make(cause._tag, cause.fiberId as unknown))
         if (stack.length === 0) {
           return [_parallel, _sequential]
         }