+++ packages/effect/src/internal/layer/circular.ts
@@ -191,12 +191,7 @@ export const parentSpan = (span: Tracer.AnySpan): Layer.Layer<Tracer.ParentSpan>
 /** @internal */
 export const span = (
   name: string,
-  options?: {
-    readonly attributes?: Record<string, unknown> | undefined
-    readonly links?: ReadonlyArray<Tracer.SpanLink> | undefined
-    readonly parent?: Tracer.AnySpan | undefined
-    readonly root?: boolean | undefined
-    readonly context?: Context.Context<never> | undefined
+  options?: Tracer.SpanOptions & {
     readonly onEnd?:
       | ((span: Tracer.Span, exit: Exit.Exit<unknown, unknown>) => Effect.Effect<void>)
       | undefined